Avoid using QPointer with QMimeData
This does not seem safe.
